acir::circuit

Module opcodes

ACIR opcodes

This module defines the core set opcodes used in ACIR.

Re-exports

pub use function_id::AcirFunctionId;

Modules

black_box_function_call 🔒

Black box functions are ACIR opcodes which rely on backends implementing support for specialized constraints. This makes certain zk-snark unfriendly computations cheaper than if they were implemented in more basic constraints.

function_id

memory_operation 🔒

Structs

BlockId

Identifier for a block of memory

InvalidInputBitSize

MemOp

Operation on a block of memory We can either write or read at an index in memory

Enums

BlackBoxFuncCall

These opcodes represent a specialized computation. Even if any computation can be done using only assert-zero opcodes, it is not always efficient. Some proving systems, can implement several computations more efficiently using techniques such as custom gates and lookup tables.

BlockType

Type for a memory block

FunctionInput

Enumeration for black box function inputs

Opcode

Defines an operation within an ACIR circuit

Functions

display_opcode 🔒

Displays an opcode, optionally using the provided return values to prefer displaying ASSERT return_value = ... when possible.