acir::circuit::opcodes

Module black_box_function_call

Black box functions are ACIR opcodes which rely on backends implementing support for specialized constraints. This makes certain zk-snark unfriendly computations cheaper than if they were implemented in more basic constraints.

Structs

InvalidInputBitSize

Enums

BlackBoxFuncCall

These opcodes represent a specialized computation. Even if any computation can be done using only assert-zero opcodes, it is not always efficient. Some proving systems, can implement several computations more efficiently using techniques such as custom gates and lookup tables.

FunctionInput

Enumeration for black box function inputs

Functions

deserialize_big_array_into_box 🔒

serialize_big_array 🔒

slice_to_string 🔒