acir::circuit
Struct Circuit

pub struct Circuit<F: AcirField> {
    pub function_name: String,
    pub current_witness_index: u32,
    pub opcodes: Vec<Opcode<F>>,
    pub private_parameters: BTreeSet<Witness>,
    pub public_parameters: PublicInputs,
    pub return_values: PublicInputs,
    pub assert_messages: Vec<(OpcodeLocation, AssertionPayload<F>)>,
}

Representation of a single ACIR circuit. The execution trace of this structure is dictated by the construction of a crate::native_types::WitnessMap

Fields
function_name: String

Name of the function represented by this circuit.

current_witness_index: u32

The current highest witness index in the circuit.

This is tracked as an optimization so that when new witness values are created, incrementing this witness results in a new unique witness index without needing to scan all opcodes to find the maximum witness index.

opcodes: Vec<Opcode<F>>

The circuit opcodes representing the relationship between witness values.

The opcodes should be further converted into a backend-specific circuit representation. When initial witness inputs are provided, these opcodes can also be used for generating an execution trace.

private_parameters: BTreeSet<Witness>

The set of private inputs to the circuit.

public_parameters: PublicInputs

The set of public inputs provided by the prover.

return_values: PublicInputs

The set of public inputs calculated within the circuit.

assert_messages: Vec<(OpcodeLocation, AssertionPayload<F>)>

Maps opcode locations to failed assertion payloads. The data in the payload is embedded in the circuit to provide useful feedback to users when a constraint in the circuit is not satisfied.

Implementations

impl<F: AcirField> Circuit<F>

pub fn circuit_arguments(&self) -> BTreeSet<Witness>

Returns all witnesses which are required to execute the circuit successfully.

pub fn public_inputs(&self) -> PublicInputs

Returns all public inputs. This includes those provided as parameters to the circuit and those computed as return values.

impl Circuit<FieldElement>

pub fn from_str(src: &str) -> Result<Self, AcirParserErrorWithSource>

Creates a Circuit object from the given string.

pub fn from_str_impl(src: &str) -> Result<Self, AcirParserErrorWithSource>