acir
Enum SerializationFormat

#[repr(u8)]
pub enum SerializationFormat {
    Msgpack = 2,
    MsgpackCompact = 3,
}

A marker byte for the serialization format.

Variants

Msgpack = 2

Msgpack with named structs.

MsgpackCompact = 3

Msgpack with tuple structs.

Implementations

impl Format

pub fn from_env() -> Result<Option<Self>, String>

Look for a NOIR_SERIALIZATION_FORMAT env var to turn on formatted serialization.

The reason we use an env var is because:

It has to be picked up in methods like Program::serialize_program_base64 where no config is available.

At the moment this is mostly for testing, to be able to commit code that can produce different formats, but only activate it once a version of bb that can handle it is released.