acir::proto::brillig
Enum IntegerBitSize

#[repr(i32)]
pub enum IntegerBitSize {
    IbsUnspecified = 0,
    IbsU1 = 1,
    IbsU8 = 8,
    IbsU16 = 16,
    IbsU32 = 32,
    IbsU64 = 64,
    IbsU128 = 128,
}

Variants

IbsUnspecified = 0

IbsU1 = 1

IbsU8 = 8

IbsU16 = 16

IbsU32 = 32

IbsU64 = 64

IbsU128 = 128

Implementations

impl IntegerBitSize

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of IntegerBitSize.

pub fn from_i32(value: i32) -> Option<IntegerBitSize>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a IntegerBitSize, or None if value is not a valid variant.

impl IntegerBitSize

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.