acir
Module serialization

Serialization formats we consider using for the bytecode and the witness stack.

Enums

Format 🔒

A marker byte for the serialization format.

Constants

FORMAT_ENV_VAR 🔒

Functions

bincode_deserialize 🔒

Deserialize a value using bincode, based on serde.

bincode_serialize 🔒

Serialize a value using bincode, based on serde.

deserialize_any_format 🔒

Deserialize any of the supported formats. Try go guess the format based on the first byte, but fall back to the legacy bincode format if anything fails.

msgpack_deserialize 🔒

Deserialize a value using MessagePack, based on serde.

msgpack_serialize 🔒

Serialize a value using MessagePack, based on serde.

serialize_with_format 🔒

serialize_with_format_from_env 🔒