acir::native_types::expression
Function display_expression

pub(crate) fn display_expression<F: AcirField>(
    expr: &Expression<F>,
    as_equal_to_zero: bool,
    return_values: Option<&PublicInputs>,
    f: &mut Formatter<'_>,
) -> Result

Displays an expression as a quadratic polynomial. If as_equal_to_zero is true, the expression is displayed as equaling zero, where it’s tried to shown as a polynomial equal to the largest witness, if possible. If the optional return_values is provided, the expression is displayed preferring to show ASSERT return_value = ... when possible.