acir::native_types
Struct Expression

pub struct Expression<F> {
    pub mul_terms: Vec<(F, Witness, Witness)>,
    pub linear_combinations: Vec<(F, Witness)>,
    pub q_c: F,
}

An expression representing a quadratic polynomial.

This struct is primarily used to express arithmetic relations between variables. It includes multiplication terms, linear combinations, and a constant term.

Addition polynomial

Unlike standard plonk constraints with fixed wire assignments (wL, wR, wO), we allow arbitrary fan-in and fan-out. This means we need a more flexible representation and we need more than wL, wR, and wO.

When looking at the quotient polynomial for the assert-zero opcode in standard plonk, you can interpret the structure in two ways:

Fan-in 2 and fan-out 1

Fan-in 1 and fan-out 2

Multiplication polynomial

If we were allow the degree of the quotient polynomial to be arbitrary, then we will need a vector of wire values.

Fields
mul_terms: Vec<(F, Witness, Witness)>

Collection of multiplication terms.

To avoid having to create intermediate variables pre-optimization We collect all of the multiplication terms in the assert-zero opcode A multiplication term is of the form q_M * wL * wR Hence this vector represents the following sum: q_M1 * wL1 * wR1 + q_M2 * wL2 * wR2 + .. +

linear_combinations: Vec<(F, Witness)>

Collection of linear terms in the expression.

Each term follows the form: q_L * w, where q_L is a coefficient and w is a witness.

q_c: F

A constant term in the expression

Implementations

impl<F: Ord> Expression<F>

fn get_max_idx(&self) -> WitnessIdx

fn get_max_term(&self, idx: &mut WitnessIdx) -> Option<Witness>

Returns the maximum witness at the provided position, and decrement the position.

This function assumes the gate is sorted

fn cmp_max(m1: Option<Witness>, m2: Option<Witness>) -> Ordering

impl<F> Expression<F>

pub fn num_mul_terms(&self) -> usize

Returns the number of multiplication terms

pub fn push_addition_term(&mut self, coefficient: F, variable: Witness)

Adds a new linear term to the Expression.

pub fn push_multiplication_term( &mut self, coefficient: F, lhs: Witness, rhs: Witness, )

Adds a new quadratic term to the Expression.

pub fn is_const(&self) -> bool

Returns true if the expression represents a constant polynomial.

Examples:

f(x,y) = x + y would return false

f(x,y) = xy would return false, the degree here is 2

f(x,y) = 5 would return true, the degree is 0

pub fn to_const(&self) -> Option<&F>

Returns a FieldElement if the expression represents a constant polynomial. Otherwise returns None.

Examples:

f(x,y) = x would return None

f(x,y) = x + 6 would return None

f(x,y) = 2*y + 6 would return None

f(x,y) = x + y would return None

f(x,y) = 5 would return FieldElement(5)

pub fn is_linear(&self) -> bool

Returns true if highest degree term in the expression is one or less.

mul_term in an expression contains degree-2 terms

linear_combinations contains degree-1 terms

Hence, it is sufficient to check that there are no mul_terms

Examples:

f(x,y) = x + y would return true

f(x,y) = xy would return false, the degree here is 2

f(x,y) = 0 would return true, the degree is 0

pub fn is_degree_one_univariate(&self) -> bool

Returns true if the expression can be seen as a degree-1 univariate polynomial

mul_terms in an expression can be univariate, however unless the coefficient is zero, it is always degree-2.

linear_combinations contains the sum of degree-1 terms, these terms do not need to contain the same variable and so it can be multivariate. However, we have thus far only checked if linear_combinations contains one term, so this method will return false, if the Expression has not been simplified.

Hence, we check in the simplest case if an expression is a degree-1 univariate, by checking if it contains no mul_terms and it contains one linear_combination term.

Examples:

f(x,y) = x would return true

f(x,y) = x + 6 would return true

f(x,y) = 2*y + 6 would return true

f(x,y) = x + y would return false

f(x,y) = x + x should return true, but we return false *** (we do not simplify)

f(x,y) = 5 would return false

pub fn sort(&mut self)

Sorts opcode in a deterministic order XXX: We can probably make this more efficient by sorting on each phase. We only care if it is deterministic

pub(crate) fn is_sorted(&self) -> bool

impl<F: AcirField> Expression<F>

pub fn from_field(q_c: F) -> Self

pub fn zero() -> Self

pub fn is_zero(&self) -> bool

pub fn one() -> Self

pub fn is_one(&self) -> bool

pub fn to_witness(&self) -> Option<Witness>

Returns a Witness if the Expression can be represented as a degree-1 univariate polynomial. Otherwise returns None.

Note that Witness is only capable of expressing polynomials of the form f(x) = x and not polynomials of the form f(x) = mx+c , so this method has extra checks to ensure that m=1 and c=0

pub fn add_mul(&self, k: F, b: &Self) -> Self

Returns self + k*b

pub fn width(&self) -> usize

Determine the width of this expression. The width meaning the number of unique witnesses needed for this expression.

impl Expression<FieldElement>

pub fn from_str(src: &str) -> Result<Self, AcirParserErrorWithSource>

Creates a Expression object from the given string.

pub fn from_str_impl(src: &str) -> Result<Self, AcirParserErrorWithSource>